adding changes

+++ src/main/java/com/astroenergy/java/astroenergyApplication/dao/AppointMentRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.astroenergy.java.astroenergyApplication.model.Appointment;
 import com.astroenergy.java.astroenergyApplication.model.UserProfile;
 
-public interface AppointMentRepo   extends CrudRepository<Appointment, Integer> {
+public interface AppointMentRepo   extends CrudRepository<Appointment, Long> {
 
 }
